Address LGTM findings

+++ libsrc/flatbufserver/FlatBufferClient.cpp
@@ -175,7 +175,7 @@ void FlatBufferClient::handleImageCommand(const hyperionnet::Image *image)
 
 		if (channelCount == 4)
 		{
-			for (int source=0, destination=0; source < width * height * sizeof(ColorRgb); source+=sizeof(ColorRgb), destination+=sizeof(ColorRgba))
+			for (int source=0, destination=0; source < width * height * static_cast<int>(sizeof(ColorRgb)); source+=sizeof(ColorRgb), destination+=sizeof(ColorRgba))
 			{
 				memmove((uint8_t*)imageRGB.memptr() + source, imageData->data() + destination, sizeof(ColorRgb));
 			}
